pyproject.toml
[tool.poetry]
name = "gifjazz"
version = "0.1.0"
description = ""
authors = ["Sergey Trofimov <truetug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
pydantic-settings = "^2.7.1"
pillow = "^11.1.0"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
dependency-injector = "^4.45.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
